pyproject.toml
[project]
name = "math-tag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "numpy>=2.4.2",
    "openai>=2.16.0",
    "pyyaml>=6.0.2",
    "regex>=2026.1.15",
    "requests>=2.32.5",
    "sympy>=1.14.0",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
]
